Added animations for score and background
Need to add for character

+++ Gravity Evader/Assets/Scripts/AppleScript.cs
@@ -8,7 +8,7 @@ public class AppleScript : MonoBehaviour {
     
 	void Start ()
     {
-        speed = 0.02f + ((float)(ScoreScript.score / 50) * (0.02f));
+        speed = 0.02f + ((ScoreScript.score / 50) * (0.02f));
     }
 	
 	void Update ()
