+++ Gravity Evader/Assets/Scripts/SilverAppleScript.cs
@@ -8,7 +8,7 @@ public class SilverAppleScript : MonoBehaviour {
 
     void Start()
     {
-        speed = 0.03f + ((float)(ScoreScript.score / 50) * (0.01f));
+        speed = 0.03f + ((ScoreScript.score / 50f) * (0.01f));
     }
 
     void Update()
